requirements.txt.txt
PyQt5
mss
Pillow
google-cloud-vision
google-cloud-translate
google-auth
keyboard